requirements.txt
bottle
jsonpickle
joblib
numpy
bios
pydot
pyyaml
pillow
tqdm
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.2+cpu
timm